+++ CHANGES.rst
@@ -40,8 +40,33 @@ Changelog
 - Use roles 'Reviewer' and 'Contributor' in meetingadvice_workflow
 - Added bypass for users having 'Manage portal' in MeetingItemWorkflowConditions in 'mayWait_advices_from', 'mayValidate' and 'mayPresent'
 
-4.1.22.2 (unreleased)
+4.1.23.4 (unreleased)
 ---------------------
+
+- In `Migrate_To_4105._cleanFTWLabels`, be sure to keep old values in case still a `PersistentList` instead removing the annotation
+- In `Migrate_To_4105._removeBrokenAnnexes`, manage parent's modification date to keep old value because removing an annex
+  will `notifyModifiedAndReindex` it's container
+
+4.1.23.3 (2020-04-30)
+---------------------
+
+- Added ram.cache for `PMCategoryVocabulary.__call__`, the vocabulary used for annex `content_category`,
+  this is useful for the `@@item_duplicate_form` that calls it many times
+- Added vocabulary `Products.PloneMeeting.Users` and using it for `person.userid` field,
+  this vocabulary displays the fullname and the userid
+
+4.1.23.2 (2020-04-29)
+---------------------
+
+- In `MeetingItem.xml`, REALLY remove the action having id `duplicate_and_keep_link`...
+
+4.1.23.1 (2020-04-29)
+---------------------
+
+- In `MeetingItem.xml`, remove the action having id `duplicate_and_keep_link`.
+
+4.1.23 (2020-04-29)
+-------------------
 
 - Added `ZLogHandler` in `Migrator.initNewHTMLFields` and in `Migrate_To_4105._cleanFTWLabels` as these steps may take some time
 - Moved `MeetingInsertingMethodsHelpMsgView` logic from `__init__` to `__call__` because errors are swallowed in `__init__`,
@@ -59,6 +84,13 @@ Changelog
     - Keep a link to original item?
     - Select annexes to keep
     - Select decision annexes to keep
+    - Annexes and decision annexes that will not be kept because using a scan_id or used annex_type is restricted to MeetingManagers
+      and current user is not a MeetingManager will be displayed greyed
+- In `vocabularies.BaseHeldPositionsVocabulary`, query only `held_positions` that are in `review_state` `active`,
+  moreover, display the `WorkflowState` viewlet on `person view` and `held_position view`
+- Fixed `showAddAnnex` and `showAddAnnexDecision` in `@@categorized-annexes`, rely on the `content_category` field vocabulary
+
+- Fix MeetingUser migration when no gender setted
 
 4.1.22.1 (2020-04-24)
 ---------------------